renderervk: C++ compatibility for tr_mesh

+++ code/renderervk/tr_mesh.c
@@ -38,7 +38,7 @@ static float ProjectRadius( float r, vec3_t location )
 		return 0;
 
 	p[0] = 0;
-	p[1] = fabs( r );
+	p[1] = fabs( (double)(r) );
 	p[2] = -dist;
 
 #if 0
